Split Tween systems in 2, one for loading and one for updating + other improvements
